extract fetchZodiacModules
