+++ deployables/app/app/routes/$avatar.$chainId/modules.ts
@@ -1,0 +1,3 @@
+export const loader = () => {
+  return []
+}
