+++ packages/modules/src/ZodiacModule.ts
@@ -1,9 +1,12 @@
 import { KnownContracts } from '@gnosis.pm/zodiac'
 
-export type SupportedModuleType =
-  | KnownContracts.DELAY
-  | KnownContracts.ROLES_V1
-  | KnownContracts.ROLES_V2
+export const SUPPORTED_ZODIAC_MODULES = [
+  KnownContracts.DELAY,
+  KnownContracts.ROLES_V1,
+  KnownContracts.ROLES_V2,
+] as const
+
+export type SupportedModuleType = (typeof SUPPORTED_ZODIAC_MODULES)[number]
 
 export type ZodiacModule = {
   moduleAddress: string
@@ -12,7 +15,7 @@ export type ZodiacModule = {
   modules?: ZodiacModule[]
 }
 
-export const ZODIAC_MODULE_NAMES = {
+export const ZODIAC_MODULE_NAMES: Record<SupportedModuleType, string> = {
   [KnownContracts.DELAY]: 'Delay',
   [KnownContracts.ROLES_V1]: 'Roles v1',
   [KnownContracts.ROLES_V2]: 'Roles v2',
