+++ packages/modules/src/index.ts
@@ -1,2 +1,3 @@
-export { ZODIAC_MODULE_NAMES } from './ZodiacModule'
+export { fetchZodiacModules } from './fetchZodiacModules'
+export { SUPPORTED_ZODIAC_MODULES, ZODIAC_MODULE_NAMES } from './ZodiacModule'
 export type { SupportedModuleType, ZodiacModule } from './ZodiacModule'
